+++ R/google_map_layer_circle.R
@@ -163,36 +163,10 @@ add_circles <- function(map,
   palette <- paletteCheck(palette)
   ## END PARAMETER CHECKS
 
+  mapObject(map, data, legend, legend_options, palette, digits, update_map_view,
+            layer_id, objArgs, list("fill_colour" = fill_colour,
+                                    "stroke_colour" = stroke_colour))
 
-  allCols <- circleColumns()
-  requiredCols <- requiredCircleColumns()
-  colourColumns <- shapeAttributes(fill_colour, stroke_colour)
-
-  shape <- createMapObject(data, allCols, objArgs)
-  pal <- createPalettes(shape, colourColumns)
-  colour_palettes <- createColourPalettes(data, pal, colourColumns, palette)
-  colours <- createColours(shape, colour_palettes)
-
-  if(length(colours) > 0){
-    shape <- replaceVariableColours(shape, colours)
-  }
-
-  ## LEGEND
-  if(any(vapply(legend, isTRUE, T))){
-    legend <- constructLegend(colour_palettes, legend)
-    if(!is.null(legend_options)){
-      legend <- addLegendOptions(legend, legend_options)
-    }
-  }
-
-  requiredDefaults <- setdiff(requiredCols, names(shape))
-  if(length(requiredDefaults) > 0){
-    shape <- addDefaults(shape, requiredDefaults, "circle")
-  }
-
-  shape <- jsonlite::toJSON(shape, digits = digits)
-
-  invoke_method(map, 'add_circles', shape, update_map_view, layer_id, legend)
 }
 
 #' @rdname clear
@@ -236,10 +210,6 @@ update_circles <- function(map, data, id,
 
   ## TODO:
   ## - is 'info_window' required, if it was included in the original add_ call?
-  ## - update legend:
-  ## --- if a legend was plotted as part of the original layer, need to
-  ## --- get the reference to it, then overwrite the containers
-  ## ---
 
   objArgs <- match.call(expand.dots = F)
 
@@ -248,7 +218,8 @@ update_circles <- function(map, data, id,
   palette <- paletteCheck(palette)
 
   allCols <- circleColumns()
-  requiredCols <- requiredCircleColumns()
+  #requiredCols <- requiredCircleColumns()
+  requiredCols <- requireColumns(match.call()[[1]])
   colourColumns <- shapeAttributes(fill_colour, stroke_colour)
 
   shape <- createMapObject(data, allCols, objArgs)